refactor: Split layermanager service into more files
Move layermanager related subscriptions from component to service
