+++ projects/hslayers/components/layer-manager/editor/layer-editor.component.ts
@@ -20,8 +20,12 @@ import {HsLayerEditorDimensionsComponent} from '../dimensions/layer-editor-dimen
 import {HsLayerEditorService} from './layer-editor.service';
 import {HsLayerEditorSublayerService} from './layer-editor-sub-layer.service';
 import {HsLayerEditorWidgetContainerService} from '../widgets/layer-editor-widget-container.service';
+import {
+  HsLayerManagerCopyLayerService,
+  HsLayerManagerService,
+} from 'hslayers-ng/shared/layer-manager';
 import {HsLayerManagerRemoveLayerDialogComponent} from '../dialogs/remove-layer-dialog.component';
-import {HsLayerManagerService} from 'hslayers-ng/shared/layer-manager';
+import {HsLayerManagerUtilsService} from 'hslayers-ng/shared/layer-manager';
 import {HsLayerUtilsService} from 'hslayers-ng/shared/utils';
 import {HsLayoutService} from 'hslayers-ng/shared/layout';
 import {HsLegendWidgetComponent} from '../widgets/legend-widget.component';
@@ -74,6 +78,8 @@ export class HsLayerEditorComponent {
     public HsEventBusService: HsEventBusService,
     public HsDialogContainerService: HsDialogContainerService,
     public hsWidgetContainerService: HsLayerEditorWidgetContainerService,
+    private hsLayerManagerUtilsService: HsLayerManagerUtilsService,
+    private hsLayerManagerCopyLayerService: HsLayerManagerCopyLayerService,
   ) {}
 
   createWidgets() {
@@ -108,7 +114,7 @@ export class HsLayerEditorComponent {
     );
     const confirmed = await dialog.waitResult();
     if (confirmed == 'yes') {
-      return this.HsLayerManagerService.saveGeoJson();
+      return this.hsLayerManagerUtilsService.saveGeoJson();
     }
   }
 
@@ -246,7 +252,7 @@ export class HsLayerEditorComponent {
     );
     const result = await dialog.waitResult();
     if (result.confirmed == 'yes') {
-      return this.HsLayerManagerService.copyLayer(result.layerTitle);
+      return this.hsLayerManagerCopyLayerService.copyLayer(result.layerTitle);
     }
   }
 }
