+++ projects/hslayers/components/layer-manager/gallery/layer-manager-gallery.component.html
@@ -42,9 +42,9 @@
             </div>
             <div (click)="toggleBasemap()">
                 <div style="height: 4.5em; width: 100%; background: white"
-                    [ngClass]="{'active': !hsLayerManagerService.data.baselayersVisible}"></div>
+                    [ngClass]="{'active': !hsLayerManagerVisiblityService.baselayersVisible}"></div>
                 <a class="dropdown-item flex-grow-1 d-flex align-items-center text-truncate"
-                    [ngClass]="{'active': !hsLayerManagerService.data.baselayersVisible}">{{'LAYERMANAGER.baseMapGallery.noBaseMap'
+                    [ngClass]="{'active': !hsLayerManagerVisiblityService.baselayersVisible}">{{'LAYERMANAGER.baseMapGallery.noBaseMap'
                     | translateHs }}</a>
             </div>
         </div>
