+++ projects/hslayers/shared/layer-manager/public-api.ts
@@ -2,3 +2,6 @@ export * from './layer-manager.service';
 export * from './layer-selector.service';
 export * from './layer-editor-vector-layer.service';
 export * from './layer-manager-metadata.service';
+export * from './layer-manager-copy-layer.service';
+export * from './layer-manager-utils.service';
+export * from './layer-manager-visiblity.service';
